Fine-tune the home window
The home window now has a title that matches the title
of the menu item that opens it. Also, the home window no
longer has an entry in the Window menu (for the menu item
that opens the home window is also in the Window menu, and
having the same window effectively listed twice is silly).